parse: dirtily resolve EventChunk from TimestampedTracePackets
This is a quick and dirty commit that resolves api::EventChunk from
itm_decode::TimestampedTracePackets. For now, only
TracePacket::{Overflow, ExceptionTrace} are converted because of the
presently small scope of the API. Only exceptions and IRQ numbers that
are mapped can be converted, but all packets are still saved to file.

XXX api::EventChunk { timestamp } is a dummy variable. Proper timestamps
have yet to be resolved.

Related to #3.
Related to #4.