trace: refactor Source structs
The two use-cases of cargo-rtic-scope (trace and replay) can be
abstracted as using two possible configurations of a source. Either the
target is traced over serial and the stream decoded into packets, or the
packets are directly deserialized from a recorded trace. The differances
here can be abstracted behind an

     Iterator<Item = Result<TimestampedTracePackets>>

which will be exploited in later commits.