refactor impl for resolv error report, general report improvements
ResolveError was only used in main, and as such had no relation with
sinks. This commit also treats ITM decode failures as errors instead of
warnings what with potietial data being lost.

Unmappable events are still considered warnings. No potential data is
lost, only a relation, which can still be assigned in the frontend.

Overflow events are also reported now. It was prerviously assumed that
the frontend would communicate this to the user, but with how much we
already report to the user, it makes no sense to omit that.