build: restore cargo build output somewhat
Somewhere along the line in either cargo itself or cargo_metadata
real-time build progress is no longer printed. Not a cause of concern,
but if there is no build cache, it will look like an invocation is
stuck. But errors are now printed as expected, which is the important
part.

Perhaps some rudimentary spinner should be added in the future?