trace: use different DWT channels/units for enter/exit events
As per #47, using a single DWT channel is less robust than using two. By
using two, a state must not be kept on the host-side, and it allows us
to catch re-entry errors.

Do we really need to use two DWT units here? We should be able to watch
both the enter and exit addresses, and when one of them writes, the ITM
package should indicate which variable was written to.