examples: add simple software task example
While we do have a dedicated repo for RTIC examples, it makes more sense
to merge those too with this repo: cargo-rtic-scope, cortem-m-rtic-trace
and relevant examples are all interlinked and must be in phase with
eachother.

However, we can't add it to the workspace yet because
per-package-target[0] is not yet stable.

[0] https://github.com/rust-lang/cargo/issues/9406