recovery: refactor, record software task dispatchers
The recovery module sees a significant overhaul in this commit. The
module is now composed of fewer types, less spaghetti, and is now more
composable.

Software task dispatchers are now also recorded. When a hardware
exception occurs we intercept the packet but treat still treat it as an
unknown event type. More apt would be to discard it, but then it would
not be saved to the replay file. Either the packet is pushed to some
EventType::Ignored or add some dispatcher-case enum for EventType::Task,
which a frontend would presumebly ignore. Either of these approaches
requires a revamp of the API, which I will have to consider on.

An API-sparing approach would be to save TracePackets to the replay file
instead of API structures, but this would break the generality of the
single source, multiple sinks implementation approach with an edge-case.