manifest: refactor with macros
This module is a bit messy and I feel like it can be improved further.
Someone must have had the same use-case of clap + Cargo.toml
metadata (or some separate TOML file), but I have yet to find a neat
solution. This commit makes the module a bit more bearable to read in
any case.