buffer packets and signal halt with channels
This commit implements packet buffering by help of a separate thread and
an unbounded channel. The separate thread is used to continuously can
source.next() so that we clear the source buffer as quickly as possible.
The crossbeam-channel API does not enable us to query the number of
buffered packets, but I suppose this is not that important. Closes #73.

The halt signal now also uses channel. This channel is a bounded channel
of zero size, effectively making it a randezvous channel: the SIGINT
handler will not exit until the SIGINT has been received in run_loop.
This seems to close #70: we no longer bog down a core to check the flag.
The performance drain now seems to be source.next().

As a side effect of using channels instead of futures::stream, we can
also close #69.