print frontend stderrs asynchronously
This commit wraps the stderr outputs of all spawned frontends in an
async-compatible struct that implements Stream::next. Instead of
flushing all stderrs after run_loop, this is now done before printing
the continuous status line. This does mean that frontend errors are only
printed after a packet has been received, but it's better than
previously.

A problem with the current implementation is that stderr choice is
biased towards the first frontend. If two frontends prints just as much
to their respective stderrs only the first's will be printed during run.
We can remedy this by returning a vector of stderr strings instead of
just the one.

Closes #33.