print eventual stderrs, status line every 100ms
Previous commit made it so stderrs overwrote the status line which was
not rendered until a new packet was received (and no error was reported
from the frontend for the same packet; never for the dummy that is). In
this commit run_loop is refactored such that we select! on the next
packet and the halt condition, and if they do not finish within 100ms,
we poll for any eventual stderrs and render the status line, and then
restart the loop.

This commit closes #79: the status line contains a trace duration
counter.