sources/tty: apply user-supplied baud rate
Before this commit, dev in

    --serial /path/to/dev

would always be configured for 115200 B/s, even if the user supplied
another rate in the manifest via tpiu_baud, e.g. 9600 B/s. A baud rate
that converts to a nix::sys::termios::BaudRate that isn't B0 must be
supplied.

Ideally we should CI this, but that requires a CI refactor so that
--serial can be specified. This is non-trivial, and we better handle #77
first.