dummy: plot worst-case enter/exits of tasks
This is a quick-and-dirty demo for plotting the states of traced RTIC
tasks over time. The used timestamp is the worst-case timestamp.
Timestamp quality and the time required to context switch between
tasks (that is, the time between a task exit and a HW task returned
action) are not rendered.