replay: on --list: print non-exhaustive header, comment if not None
The header does not describe the optional comment field because we do
not yet align the printout after the length of the file name.

The best approach would be to collect all file metadata and find the
longest file name length from these, but that requires fallible
iterators which I do not have time to investigate currently.